+++ mercury/src/main/java/org/broadinstitute/gpinformatics/mercury/entity/workflow/LabBatchStartingVessel.java
@@ -15,6 +15,7 @@ import javax.persistence.Id;
 import javax.persistence.ManyToOne;
 import javax.persistence.SequenceGenerator;
 import javax.persistence.Table;
+import java.math.BigDecimal;
 
 @Entity
 @Audited
@@ -34,7 +35,7 @@ public class LabBatchStartingVessel {
     private LabVessel labVessel;
 
     @Column
-    private Float concentration;
+    private BigDecimal concentration;
 
     @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.PERSIST)
     private LabVessel dilutionVessel;
@@ -47,7 +48,7 @@ public class LabBatchStartingVessel {
     }
 
     public LabBatchStartingVessel(@Nonnull LabVessel labVessel, @Nonnull LabBatch labBatch,
-                                  @Nullable Float concentration) {
+                                  @Nullable BigDecimal concentration) {
         this.labVessel = labVessel;
         this.labBatch = labBatch;
         this.concentration = concentration;
@@ -77,11 +78,11 @@ public class LabBatchStartingVessel {
         this.labVessel = vessel;
     }
 
-    public float getConcentration() {
+    public BigDecimal getConcentration() {
         return concentration;
     }
 
-    public void setConcentration(float concentration) {
+    public void setConcentration(BigDecimal concentration) {
         this.concentration = concentration;
     }
