+++ mercury/src/test/java/org/broadinstitute/gpinformatics/mercury/boundary/lims/LimsQueryResourceUnitTest.java
@@ -19,6 +19,7 @@ import org.testng.Assert;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
@@ -397,7 +398,7 @@ public class LimsQueryResourceUnitTest {
     @Test(groups = DATABASE_FREE, enabled = true)
     public void testFetchIlluminaSeqTemplate() {
         SequencingTemplateLaneType laneType =
-                LimsQueryObjectFactory.createSequencingTemplateLaneType("LANE_1234", 33.333f, "LOADING_VESSEL_1234",
+                LimsQueryObjectFactory.createSequencingTemplateLaneType("LANE_1234", BigDecimal.valueOf(33.333f), "LOADING_VESSEL_1234",
                         "LOADING_VESSEL_1234");
         SequencingTemplateType template =
                 LimsQueryObjectFactory
@@ -418,7 +419,7 @@ public class LimsQueryResourceUnitTest {
         Assert.assertTrue(result.isPairedRun());
         Assert.assertEquals(result.getLanes().size(), 1);
         SequencingTemplateLaneType laneOne = result.getLanes().get(0);
-        Assert.assertEquals(laneOne.getLoadingConcentration(), 33.333f);
+        Assert.assertEquals(laneOne.getLoadingConcentration(), BigDecimal.valueOf(33.333f));
         Assert.assertEquals(laneOne.getLoadingVesselLabel(), "LOADING_VESSEL_1234");
         Assert.assertEquals(laneOne.getLaneName(), "LANE_1234");
         verifyAll();
